Update bStats for Sponge to utilize Sponge's built-in config:
Now compliant with guidelines both present and soon-to-be-changed.

Utilizes Sponge's metrics manager to determine if each plugin should
send stats. If no plugin can, it does not send stats.

Replaces save-classname-to-file approach with use of services API.
Introduced new Metrics interface, which conveniently can't exist
for older bStats because using the package wasn't allowed.
Renamed Metrics to Metrics2 for classname-based versioning.
Renamed MetricsLite to MetricsLite2 for the same.

Now enforcing that the classes NOT be relocated. Technically the
individual instances (Metrics2, MetricsLite2) could be, but may as
well force them to not move alongside the Metrics interface.

Separate bStats 'enabled' config value is left, but a new comment is
added which explains where to look for the future. The
guidelines-required default value of disabled is added in case the
user first loads a new bStats using plugin followed by an older one.

As long as a bStats 2 instance is loaded, it will take over for any
bStats 1 instance that loads, ensuring those plugins also now follow
the guidelines and, more importantly, ensuring data is only sent
once instead of new and old instances sending data separately,
confusing bStats. Provides mechanisms for easier replacement of
itself should a bStats 3 be created at some point in the future.

Example added to BOTH projects.

Enjoy!